# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    CreateStudentRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        externalId:
          type: string
      required:
        - email
        - name
    BulkStudentRequest:
      type: object
      properties:
        csv:
          type: string
          description: CSV data with columns email,name,external_id
      required:
        - csv
    BulkResult:
      type: object
      properties:
        created:
          type: integer
        errors:
          type: array
          items:
            type: string
      required:
        - created
        - errors
    StudentInfo:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        externalId:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - isActive
    Book:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
          nullable: true
        isbn:
          type: string
          nullable: true
        language:
          type: string
        level:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - language
        - isActive
        - createdAt
        - updatedAt
...
    BookAssignment:
      type: object
      properties:
        id:
          type: integer
        bookId:
          type: integer
        studentId:
          type: integer
        institutionId:
          type: integer
        assignedBy:
          type: integer
          nullable: true
        assignedAt:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
          nullable: true
        isActive:
          type: boolean
        book:
          $ref: "#/components/schemas/Book"
        institution:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
      required:
        - id
        - bookId
        - studentId
        - institutionId
        - assignedAt
        - isActive
    CreateAssignmentRequest:
      type: object
      properties:
        bookId:
          type: integer
        studentId:
          type: integer
        dueDate:
          type: string
          format: date-time
      required:
        - bookId
        - studentId
    BulkAssignmentRequest:
      type: object
      properties:
        bookId:
          type: integer
        studentIds:
          type: array
          items:
            type: integer
        dueDate:
          type: string
          format: date-time
      required:
        - bookId
        - studentIds
    BulkAssignmentResult:
      type: object
      properties:
        created:
          type: integer
        skipped:
          type: integer
      required:
        - created
        - skipped
    BulkUnassignRequest:
      type: object
      properties:
        assignmentIds:
          type: array
          items:
            type: integer
      required:
        - assignmentIds
    BulkUnassignResult:
      type: object
      properties:
        deactivated:
          type: integer
        requested:
          type: integer
      required:
        - deactivated
        - requested
...
    ReadingProgress:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        bookId:
          type: integer
        assignmentId:
          type: integer
          nullable: true
        pagesRead:
          type: integer
        totalPages:
          type: integer
          nullable: true
        completionPercent:
          type: integer
        lastReadAt:
          type: string
          format: date-time
          nullable: true
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        student:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            email:
              type: string
        book:
          type: object
          properties:
            id:
              type: integer
            title:
              type: string
            author:
              type: string
              nullable: true
      required:
        - id
        - studentId
        - bookId
        - pagesRead
        - completionPercent
        - startedAt
        - createdAt
        - updatedAt
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
paths:
...
  /operator/students:
    get:
      operationId: listStudents
      tags: [operator]
      summary: List students in operator's institution
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentInfo"
    post:
      operationId: registerStudent
      tags: [operator]
      summary: Register a student
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentRequest"
      responses:
        "201":
          description: Student registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentInfo"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /operator/students/bulk:
    post:
      operationId: bulkRegisterStudents
      tags: [operator]
      summary: Bulk register students via CSV
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkStudentRequest"
      responses:
        "200":
          description: Bulk registration result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkResult"
  /operator/assignments:
    post:
      operationId: createAssignment
      tags: [operator]
      summary: Assign a book to a student
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignmentRequest"
      responses:
        "201":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookAssignment"
  /operator/assignments/bulk:
    post:
      operationId: bulkCreateAssignments
      tags: [operator]
      summary: Assign a book to multiple students
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkAssignmentRequest"
      responses:
        "201":
          description: Bulk assignment result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkAssignmentResult"
  /operator/assignments/{id}:
    delete:
      operationId: deleteAssignment
      tags: [operator]
      summary: Deactivate a book assignment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assignment deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /operator/assignments/bulk-unassign:
    post:
      operationId: bulkUnassignBooks
      tags: [operator]
      summary: Bulk deactivate book assignments
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkUnassignRequest"
      responses:
        "200":
          description: Bulk unassign result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkUnassignResult"
  /operator/progress:
    get:
      operationId: getReadingProgress
      tags: [operator]
      summary: Get reading progress for students
      security:
        - bearerAuth: []
      parameters:
        - name: studentId
          in: query
          required: false
          schema:
            type: integer
          description: Filter by student ID
      responses:
        "200":
          description: Reading progress list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReadingProgress"
...